fs-resize: Add configurable game partition for Windows users
Previously, all remaining space on the system disk was allocated to the
storage partition. This approach may not be ideal for Windows users who
wish to add games, as it doesn't provide a dedicated Windows known partition.
Although Windows users can utilize Samba and USB methods for transferring
game data, these methods tend to be slower compared to direct copying via
a card reader, especially for large game files.

This commit introduces the ability to create a configurable "GAMES"
partition. Users can now specify the desired size of the storage partition
by creating a file named `resize_storage_xG` in the `/flash` partition
before the first system startup (after flashing the image to the disk).
The `fs-resize` script will then:

1. Resize the storage partition to end at `xG`.
2. Create a new "GAMES" partition using the remaining space on the disk.

If the `resize_storage_xG` file is not present, the `fs-resize` script
will default to allocating all available space to the storage partition as
before.

Example:
- To reserve storage partition ends with 50G and use the remaining space
for the GAMES partition, create a file named `resize_storage_50G` in
the `/flash` partition.

Signed-off-by: kk <405437445@qq.com>